+++ src/components/header/navarItems.ts
@@ -2,7 +2,6 @@ import {
   MdInfoOutline,
   MdLogin,
   MdLogout,
-  MdOutlineArrowRight,
   MdPerson,
   MdPersonAdd,
   MdSendToMobile,
@@ -60,37 +59,4 @@ const menu = [
   },
 ];
 
-const products = [
-  {
-    id: 1,
-    title: '구글기프트카드',
-    to: '/shop/categories/구글',
-    icon: MdOutlineArrowRight,
-  },
-  {
-    id: 2,
-    title: '넥슨카드',
-    to: '/shop/categories/넥슨',
-    icon: MdOutlineArrowRight,
-  },
-  {
-    id: 3,
-    title: '해피머니',
-    to: '/shop/categories/해피',
-    icon: MdOutlineArrowRight,
-  },
-  {
-    id: 4,
-    title: '컬쳐랜드상품권',
-    to: '/shop/categories/컬쳐',
-    icon: MdOutlineArrowRight,
-  },
-  {
-    id: 5,
-    title: '문화상품권',
-    to: '/shop/categories/문화',
-    icon: MdOutlineArrowRight,
-  },
-];
-
-export { authenticated, unauthenticated, menu, products };
+export { authenticated, unauthenticated, menu };
