+++ src/components/header/products.ts
@@ -1,0 +1,36 @@
+import { MdOutlineArrowRight } from 'react-icons/md';
+
+const products = [
+  {
+    id: 1,
+    title: '구글기프트카드',
+    to: '/shop/categories/구글',
+    icon: MdOutlineArrowRight,
+  },
+  {
+    id: 2,
+    title: '넥슨카드',
+    to: '/shop/categories/넥슨',
+    icon: MdOutlineArrowRight,
+  },
+  {
+    id: 3,
+    title: '해피머니',
+    to: '/shop/categories/해피',
+    icon: MdOutlineArrowRight,
+  },
+  {
+    id: 4,
+    title: '컬쳐랜드상품권',
+    to: '/shop/categories/컬쳐',
+    icon: MdOutlineArrowRight,
+  },
+  {
+    id: 5,
+    title: '문화상품권',
+    to: '/shop/categories/문화',
+    icon: MdOutlineArrowRight,
+  },
+];
+
+export { products };
